Added a store to keep track of the people attending
