+++ package.json
@@ -14,6 +14,7 @@
     "marked": "^5.1.0",
     "marked-gfm-heading-id": "^3.0.4",
     "marked-mangle": "^1.0.1",
+    "pinia": "^2.1.4",
     "vue": "^3.3.4",
     "vuetify": "^3.3.4",
     "zero-md": "^2.4.1"
