+++ src/main.js
@@ -1,10 +1,20 @@
 import "@/assets/main.css"
 
 import { createApp } from "vue"
+import { createPinia } from "pinia"
 import App from "@/App.vue"
 
-// Vuetify
+// Plugins
 import Vuetify from "@/plugins/vuetify.js"
 
-createApp(App).use(Vuetify).mount("#app")
+const pinia = createPinia()
+const app = createApp(App)
+
+// Use pinia stores
+app.use(pinia)
+
+// Plugins to use
+app.use(Vuetify)
+
+app.mount("#app")
